+++ TickeTac/Models/Event.cs
@@ -8,7 +8,7 @@ namespace TickeTac.Models
     public class Event
     {
         [Key]
-        public uint Id { get; set; }
+        public UInt16 Id { get; set; }
 
         [Display(Name = "Nome do evento")]
         [Required(ErrorMessage = "O campo não pode ser vazio!")]
